requirements.txt
pillow
python-dotenv